refactor(LottoTicketMachine): 서비스 레이어 삭제하고 로또 발권기 도메인을 생성하기
- WHY: 기존에 있던 서비스 계층을 삭제하고 로또 발권기 도메인 클래스를 생성하였다. 기존 서비스 클래스에서 로또를 발행하는 메서드 이외의 메서드는 도메인 클래스의 메서드를 전달하거나 유효성을 검사하는 것 뿐이었기에 불필요하다고 판단했다.
