+++ src/main/java/lotto/configuration/AppConfig.java
@@ -1,7 +1,6 @@
 package lotto.configuration;
 
 import lotto.controller.LottoController;
-import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
@@ -14,12 +13,8 @@ public class AppConfig {
         return new OutputView();
     }
 
-    private LottoService lottoService() {
-        return new LottoService();
-    }
-
     public LottoController lottoController() {
-        return new LottoController(inputView(), outputView(), lottoService());
+        return new LottoController(inputView(), outputView());
     }
 
 }
