+++ src/main/java/lotto/domain/LottoTicketMachine.java
@@ -1,21 +1,17 @@
-package lotto.service;
+package lotto.domain;
 
 import static lotto.constants.Condition.MAXIMUM_VALUE;
 import static lotto.constants.Condition.MINIMUM_VALUE;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
-import lotto.domain.BonusNumber;
 import lotto.domain.IssuedLottos;
 import lotto.domain.Lotto;
-import lotto.domain.Profit;
 import lotto.domain.PurchaseNumber;
-import lotto.domain.Rank;
-import lotto.utility.validation.WinningNumberChecker;
 import lotto.utility.vo.request.PurchaseAmountRequest;
 
-public class LottoService {
-    public IssuedLottos generateLottos(PurchaseAmountRequest purchaseAmount) {
+public class LottoTicketMachine {
+    static public IssuedLottos generateLottos(PurchaseAmountRequest purchaseAmount) {
         PurchaseNumber purchaseNumber = purchaseAmount.calculatePurchaseNumber();
 
         Integer count = 0;
@@ -28,17 +24,5 @@ public class LottoService {
         }
 
         return new IssuedLottos(lottoTickets);
-    }
-
-    public List<Rank> determineWinning(
-        IssuedLottos issuedLottos,
-        Lotto winningLotto,
-        BonusNumber bonusNumber) {
-        WinningNumberChecker.validate(winningLotto, bonusNumber);
-        return issuedLottos.determineRanks(winningLotto, bonusNumber);
-    }
-
-    public Profit createProfit(List<Rank> ranks) {
-        return new Profit(ranks);
     }
 }
